Refactor File/Folder structure following sazim principles.
- Create Home folder containing Home.jsx.
- Create HomeConstants.js and HomeHelpers.js in Home folder.
- Add folder for each components in Components folder.
