+++ src/pages/HomePage/Components/Bin/Bin.jsx
@@ -4,12 +4,12 @@ import { IconTrashFilled } from "@tabler/icons-react";
 const Bin = ({ trash, handleRetrieveAllClick, handleEmptyBinClick }) => {
   return (
     <>
-      <Group position="right" my={"xs"}>
+      <Group position="right" mt={"md"}>
         {trash.length !== 0 ? (
           <Menu width={200} shadow="md" withArrow position="left" offset={0}>
             <Menu.Target>
-              <ActionIcon size={"4rem"} color="red.4" variant="transparent">
-                <IconTrashFilled stroke="1px" size={"3rem"} />
+              <ActionIcon size={"3rem"} color="red.4" variant="transparent">
+                <IconTrashFilled stroke="1px" size={"2rem"} />
               </ActionIcon>
             </Menu.Target>
