+++ src/pages/HomePage/HomePageHelpers.js
